Remove "name()" from StateBuilder interface
We don't currently use the StateBuilder name for any
reason and it creates more boilerplate when creating
stateful apps. This removes name() from the interface
and updates Python and Go APIs as well as example
apps and book examples accordingly. Instead of the
StateBuilder name, we now use the state partition
name, which would do the trick if we ever end up
needing it for logging purposes.

Closes #1809